requirements.txt
# 基础依赖
numpy>=1.19.0
scipy>=1.7.0
matplotlib>=3.3.0
seaborn>=0.11.0

# 深度学习框架
torch>=1.9.0

# 强化学习环境
gym>=0.19.0

# 可视化依赖
imageio>=2.9.0
imageio-ffmpeg>=0.4.5
moviepy>=1.0.3
Pillow>=8.0.0
tqdm>=4.62.0

# 配置和测试
pyyaml>=5.4.1
pytest>=6.2.5 